HotFix 1.8 Added routers for pages
