+++ src/components/Layout/MainLayout.css
@@ -1,0 +1,18 @@
+header {
+    position: fixed;
+    width: 100vw;
+    height: 58px;
+    background-color: #40BA21;
+    box-shadow: 0 2px 10px 0 rgba(30, 30, 30, 0.25);
+    z-index: 100;
+}
+
+
+.logoText {
+    position: absolute;
+    color: #ECECEC;
+    font-family: Audiowide, serif;
+    font-size: 35px;
+    left: 23px;
+    top: 6px;
+}
