+++ src/pages/TakePhoto/TakePhoto.css
@@ -1,16 +1,12 @@
-body{
-    background-color: #0C0C0C;
-}
 
-.camera{
-    position: absolute;
+.camera {
     display: flex;
     justify-content: center;
     align-items: center;
-    top: 121px;
+    margin-top: 300px;
 }
 
-.camera_shot{
+.camera_shot {
     position: absolute;
     display: flex;
     justify-content: center;
@@ -19,7 +15,7 @@ body{
     top: 121px;
 }
 
-.button_take_photo{
+.button_take_photo {
     position: absolute;
     display: flex;
     justify-content: center;
@@ -28,11 +24,11 @@ body{
     left: 145px;
 }
 
-.icon_photo{
+.icon_photo {
     position: absolute;
 }
 
-.button_retake_photo{
+.button_retake_photo {
     color: #FFFFFF;
     display: flex;
     justify-content: center;
@@ -42,7 +38,7 @@ body{
     left: 130px;
 }
 
-.button_retake_photo svg{
+.button_retake_photo svg {
     width: 19px;
     height: 15px;
     flex-shrink: 0;
@@ -50,7 +46,7 @@ body{
     margin-top: 1px;
 }
 
-.button_submit_photo{
+.button_submit_photo {
     color: #FFFFFF;
     border-radius: 5px;
     background: #40BA21;
